+++ examples/Generic/Ethernet/Ethernet_NTPClient_Basic_STM32/defines.h
@@ -38,7 +38,7 @@
 // either use W5x00 with EthernetXYZ library
 // or ENC28J60 with EthernetENC library
 //#define USE_BUILTIN_ETHERNET    true
-//#define USE_BUILTIN_ETHERNET    false
+#define USE_BUILTIN_ETHERNET    false
 
 //#define USE_UIP_ETHERNET        true
 #define USE_UIP_ETHERNET        false