+++ src/NTPClient_Generic.h
@@ -8,6 +8,7 @@
   d) ESP8266/ESP32-AT-command WiFi
   e) WT32_ETH01 (ESP32 + LAN8720)
   f) RTL8720DN
+  g) Portenta_H7
 
   Based on and modified from Arduino NTPClient Library (https://github.com/arduino-libraries/NTPClient)
   to support other boards such as ESP8266/ESP32, SAMD21, SAMD51, Adafruit's nRF52 boards, SAM DUE, RTL8720DN, etc.
@@ -17,7 +18,7 @@
 
   Built by Khoi Hoang https://github.com/khoih-prog/NTPClient_Generic
   Licensed under MIT license
-  Version: 3.5.2
+  Version: 3.6.0
 
   Version Modified By  Date      Comments
   ------- -----------  ---------- -----------
@@ -29,6 +30,7 @@
   3.5.0   K Hoang      10/08/2021 Add support to Ameba Realtek RTL8720DN, RTL8722DM and RTL8722CSM
   3.5.1   K Hoang      12/10/2021 Update `platform.ini` and `library.json`
   3.5.2   K Hoang      01/12/2021 Auto detect ESP32 core version. Fix bug in examples for WT32_ETH01
+  3.6.0   K Hoang      08/01/2022 Fix `multiple-definitions` linker error. Add support to Portenta_H7
  *****************************************************************************************************************************/
  
 #pragma once
@@ -37,7 +39,13 @@
 #ifndef NTPCLIENT_GENERIC_H
 #define NTPCLIENT_GENERIC_H
 
-#define NTPCLIENT_GENERIC_VERSION      "NTPClient_Generic v3.5.2"
+#define NTPCLIENT_GENERIC_VERSION             "NTPClient_Generic v3.6.0"
+
+#define NTPCLIENT_GENERIC_VERSION_MAJOR       3
+#define NTPCLIENT_GENERIC_VERSION_MINOR       6
+#define NTPCLIENT_GENERIC_VERSION_PATCH       0
+
+#define NTPCLIENT_GENERIC_VERSION_INT        3006000
 
 #include "Arduino.h"
 
@@ -115,7 +123,7 @@ class NTPClient
 
        @param poolServerName
     */
-    String getPoolServerName(void)
+    String getPoolServerName()
     {
       return String(this->_poolServerName);
     }
@@ -137,7 +145,7 @@ class NTPClient
 
        @param poolServerIP
     */
-    IPAddress getPoolServerIP(void)
+    IPAddress getPoolServerIP()
     {
       return (this->_poolServerIP);
     }
@@ -398,7 +406,5 @@ class NTPClient
       this->_udpSetup = false;
     }
 };
-
-#include "NTPClient_Generic_Impl.h"
 
 #endif    // NTPCLIENT_GENERIC_H