+++ src/NTPClient_Generic_Debug.h
@@ -8,6 +8,7 @@
   d) ESP8266/ESP32-AT-command WiFi
   e) WT32_ETH01 (ESP32 + LAN8720)
   f) RTL8720DN
+  g) Portenta_H7
 
   Based on and modified from Arduino NTPClient Library (https://github.com/arduino-libraries/NTPClient)
   to support other boards such as ESP8266/ESP32, SAMD21, SAMD51, Adafruit's nRF52 boards, SAM DUE, RTL8720DN, etc.
@@ -17,7 +18,7 @@
 
   Built by Khoi Hoang https://github.com/khoih-prog/NTPClient_Generic
   Licensed under MIT license
-  Version: 3.5.2
+  Version: 3.6.0
 
   Version Modified By  Date      Comments
   ------- -----------  ---------- -----------
@@ -29,6 +30,7 @@
   3.5.0   K Hoang      10/08/2021 Add support to Ameba Realtek RTL8720DN, RTL8722DM and RTL8722CSM
   3.5.1   K Hoang      12/10/2021 Update `platform.ini` and `library.json`
   3.5.2   K Hoang      01/12/2021 Auto detect ESP32 core version. Fix bug in examples for WT32_ETH01
+  3.6.0   K Hoang      08/01/2022 Fix `multiple-definitions` linker error. Add support to Portenta_H7
  *****************************************************************************************************************************/
 
 #pragma once