+++ src/NTPClient_Generic_Impl.h
@@ -8,6 +8,7 @@
   d) ESP8266/ESP32-AT-command WiFi
   e) WT32_ETH01 (ESP32 + LAN8720)
   f) RTL8720DN
+  g) Portenta_H7
 
   Based on and modified from Arduino NTPClient Library (https://github.com/arduino-libraries/NTPClient)
   to support other boards such as ESP8266/ESP32, SAMD21, SAMD51, Adafruit's nRF52 boards, SAM DUE, RTL8720DN, etc.
@@ -17,7 +18,7 @@
 
   Built by Khoi Hoang https://github.com/khoih-prog/NTPClient_Generic
   Licensed under MIT license
-  Version: 3.5.2
+  Version: 3.6.0
 
   Version Modified By  Date      Comments
   ------- -----------  ---------- -----------
@@ -29,6 +30,7 @@
   3.5.0   K Hoang      10/08/2021 Add support to Ameba Realtek RTL8720DN, RTL8722DM and RTL8722CSM
   3.5.1   K Hoang      12/10/2021 Update `platform.ini` and `library.json`
   3.5.2   K Hoang      01/12/2021 Auto detect ESP32 core version. Fix bug in examples for WT32_ETH01
+  3.6.0   K Hoang      08/01/2022 Fix `multiple-definitions` linker error. Add support to Portenta_H7
  *****************************************************************************************************************************/
 
 #pragma once
@@ -37,11 +39,15 @@
 #ifndef NTPCLIENT_GENERIC_IMPL_H
 #define NTPCLIENT_GENERIC_IMPL_H
 
+/////////////////////////////
+
 NTPClient::NTPClient(UDP& udp, long timeOffset)
 {
   this->_udp            = &udp;
   this->_timeOffset     = timeOffset;
 }
+
+/////////////////////////////
 
 NTPClient::NTPClient(UDP& udp, const char* poolServerName, long timeOffset, unsigned long updateInterval)
 {
@@ -50,6 +56,8 @@ NTPClient::NTPClient(UDP& udp, const char* poolServerName, long timeOffset, unsi
   this->_poolServerName = poolServerName;
   this->_updateInterval = updateInterval;
 }
+
+/////////////////////////////
 
 NTPClient::NTPClient(UDP& udp, IPAddress poolServerIP, long timeOffset, unsigned long updateInterval)
 {
@@ -60,6 +68,8 @@ NTPClient::NTPClient(UDP& udp, IPAddress poolServerIP, long timeOffset, unsigned
   this->_updateInterval = updateInterval;
 }
 
+/////////////////////////////
+
 void NTPClient::begin(int port)
 {
   this->_port = port;
@@ -68,6 +78,8 @@ void NTPClient::begin(int port)
 
   this->_udpSetup = true;
 }
+
+/////////////////////////////
 
 // Perform some validity checks on the packet
 //  https://datatracker.ietf.org/doc/html/rfc4330#section-4
@@ -98,6 +110,8 @@ static bool isValid(byte const *ntpPacket)
     ((refTimeInt != 0) || (refTimeFrac != 0))
   );
 }
+
+/////////////////////////////
 
 bool NTPClient::checkResponse()
 {
@@ -138,6 +152,8 @@ bool NTPClient::checkResponse()
   return false;
 }
 
+/////////////////////////////
+
 bool NTPClient::forceUpdate()
 {
   NTP_LOGDEBUG("forceUpdate from NTP Server");
@@ -165,6 +181,8 @@ bool NTPClient::forceUpdate()
 
   return true;
 }
+
+/////////////////////////////
 
 bool NTPClient::update()
 {
@@ -207,6 +225,8 @@ unsigned long long NTPClient::getUTCEpochMillis()
   return epoch;
 }
 
+/////////////////////////////
+
 unsigned long long NTPClient::getEpochMillis() 
 {
   unsigned long long epoch;
@@ -232,15 +252,21 @@ String NTPClient::createFormattedTime(unsigned long rawTime) const
   return hoursStr + ":" + minuteStr + ":" + secondStr;
 }
 
+/////////////////////////////
+
 String NTPClient::getFormattedTime() const 
 {
   return createFormattedTime(this->getEpochTime());
 }
 
+/////////////////////////////
+
 String NTPClient::getFormattedUTCTime() const 
 {
   return createFormattedTime(this->getUTCEpochTime());
 }
+
+/////////////////////////////
 
 String NTPClient::getFormattedDateTime() const 
 {
@@ -258,6 +284,8 @@ String NTPClient::getFormattedDateTime() const
   
   return String(buf);
 }
+
+/////////////////////////////
 
 String NTPClient::getFormattedUTCDateTime() const 
 {